Replace movement and collision with resolv library
Some parts still need to be written because the current movement steps
before it checks for collisions but it should be the other way around.
At least so far we have achieved the previous game experience using this
library and also getting rid of the magic numbers for square size in
Draw() because the resolve.Object knows the height.