Stop the zombies from walking over each other
Before the zombies move check to see if it would bump into another
zombie if it moved there and if so, stay where it is until it is able to
move there.  To be able to distinguish from bumping into zombies and the
player character they have all received the "mob" tag and the collision
response logic filters only for that tag.  Because the grid size is
slightly bigger than the zombie size they stop _before_ they collide and
keep some space between each other.  This makes them seem to move much
more organically.