Replace tile walls with intgrid walls
The intgrid is set up in ldtk for walls to be able to drop edges,
corners, inner corners and single-standing tiles.  The existing walls
were replaced with empty sand on the tilemap and then redrawn using the
intgrid so that they can be referenced as walls in the game code.