Remove a zombie every time you shoot
This is a placeholder for actual shooting logic, zombies need states to
handle death animation and should only be removed at the end of the
animation.  The logic for which zombie was in the line of fire and in
range is also still missing, but should be possible using resolv.