Only kill the nearest zombie instead of all the zombies in line of fire
If the zombies stand behind each other and you shoot, it'll shoot all of
them.  This crashes the game during cleanup because after removing the
first one, the size of the array has changed and by the time it removes
the second one you hit an array out of range error.

Instead of optimising this to remove more zombies in a smarter way the
game rule has been changed:  If they're standing in a line, you will
only hit the first zombie.  Fortunately this is very easy to implement
because the CellsInLine function traverses the cells in order, so you
just need to return early after hitting the first one.