Get the player+zombie dimensions from their sprites
The zombies had a hardcoded collision object size of 16x16 pixels and
the player 20x20.  Instead of these hardcoded values we now get the
width (W) and heigh (H) of the first frame of the sprite animation.
Assuming all frames are the same size this should be good.

This also solves the problem that the player would often collide with
walls from too far away on his bottom and right side, now his collision
box matches the size of his sprite so the collisions, especially with
walls, feel much more accurate.