Hold shift to sprint
In the movement controls code, before the movement keys are handled,
check to see if Shift is being held, in which case set the new Sprinting
field to true.  This field is set to false at the start of every Update,
for lack of a better idea for managing this state.  In the MoveForward
function multiply movement speed by the sprint factor if the Player is
Sprinting.

The sprint factor is currently a little more than twice the normal
speed, it feels fast enough to make getting around the map less
annoying, without being so fast as to feel totally unrealistic.