Randomise zombies' walking speed a little bit
Instead of using the constant defined zombieSpeed this is now a base
speed.  When we construct new zombies we assign them a random speed
value in the range of zombieSpeed+0.0 to zombieSpeed+1.0.  This makes
them look less robotic because they're trundling at slightly different
speeds and also seems to have solved the weird effect of them walking in
neat queued lines, because the ones at the back can't keep up and fall
behind, so it ends up looking much more like an organic swarm.