Calculate centre of player from player object
So far this was hardcode as 8x8 because I was in a hurry and new that
the player sprites was 16x16.  The player object's resolv object already
contains width (W) and height (H) data that we set when we construct the
object so instead we can take half of that and it'll work even if the
size of the player sprite changes later on.