Define a Debugger interface for onscreen debugging
Based on the pattern from the standard library's http handlers there is
now a Debugger interface that allows any type that has a Debug method to
debug Game data onto the screen.  Because these are expected to be
mostly short callback functions there's also a default DebugFunc
implementation, much like the http.HandlerFunc, that takes any function
with a matching signature and converts it into a Debugger.

The first demonstration of this is the DebugText function which prints
the general text info to the screen, FPS, player coordinates etc.