Add Debuggers type for handling and running multiple debuggers
This is just a slice of Debugger, it has 2 methods Add and Debug.  Add
appends a new a Debugger to the internal slice, it's just a more
convenient syntax than having to write append twice.  Debug iterates
over all the Debugger-s in the slice and calls their Debug functions
with the same arguments.  Because the arguments match, Debuggers itself
also implements the Debugger interface and can be called anywhere where
previously any of its children could have been.  This seems very cool
but I'm not sure if it is a Good Thing.