Get number of zombies to spawn from the map data
The zombie entity on the map has an "initial" property that says how
many zombies it should spawn initially.  Get this number and spawn that
many.  The problem is that if you put them all in the same place then
they can't move because they're colliding.

I didn't bother coming up with a clever distribution algorithm, instead
just wrote a list of tiles spiralling around the centre point and pick
starting positions from that sequentially.  It seems like a really
stupid solution, I don't like it at all, but it works for now.