Add game over state and death screen
When you die, instead of returning an error and closing the game, change
the game global state to the new gameOver state.  This state still
handles controls so you can press F to get out of fullscreen and Q to
quit but returns before handling the rest of the normal game code.  For
draw it also returns early, drawing a death message onto a black screen
using debug text.