Implement run-time const overrides
Technically the first step was to make the consts vars, so not really
consts any more, even though their purpose is the same.  Copy-paste all
the values to an ini file, copy paste their types to an Apply function
and call it in main.  This code is quite copy-pasty but at least it
works.