Use tile grid size instead of layer grid size when cutting tiles
The renderer assumes that the tile size matches the grid size when
cutting tiles from the tileset to be rendered.  We have a tileset with
96x96 tiles (large tree canopies) that are aligned on a 32x32 layer grid
to match 32x32 tree trunks on the ground layer.  The previous
implementation cut out the top-left 32x32 corner of the 96x96 tiles
instead of drawing the whole thing because it was using the layer grid
size.  Changing that to the layer's associated tileset's tileset grid
size solved the problem and the whole 96x96 pixel tile is drawn.