Draw Treetops layer in the foreground over the entities
This requires a separate image, just like bg/background but for
fg/foreground.  After rendering the tiles in NewGame, make an exception
for this layer and draw it to fg instea of bg.  Then during draw, draw
this fg image after all the other things, players, zombies etc. have
already been drawn so that you can walk underneath the treetops.