Reload the gun every 7 shots
This is for a bit of realism but also to add tension if you have to
fight a lot of zombies, the pause can count a lot.  The value 7 has been
set as a constant playerAmmoClipMax.  The current amount of ammunition
is tracked with Player.Ammo.  If, after the shoot animation, the ammo is
at 0, instead of going to idle it will go to the reload animation.  You
can't shoot while reloading is in progress.  Future versions might
punish spam-clicking more by playing a dry shot sound and restarting the
reload animation if the player tries to shoot while reloading.