Capture cursor in the game window
This eliminates the need for you to take care to keep the cursor inside
the game window.  Also because it makes the cursor invisible it makes it
a bit harder to aim because you can't just click on the zombies, instead
you need to look where the player's facing.  This might make the need
for implementing some kind of sight/crosshair a bit more urgent.