Add a start screen for when the game starts up
This makes so that instead of jumping straight into the game you are
fist shown a start screen with text saying "press space to start" it
does nothing until you press space, then it changes the state to game
running.  This introduces the general concept of Screens which will be
the beginning of a refactor to replace the if statements currently in
Update/Draw for different states with a delegation to the Screens to
handle their own Update/Draw and state changes between them.