Add collision detection for checkpoints (POC)
This adds all the checkpoint/waypoint entities to the collision space
using the location from the map and size from their image file.  It also
assumes that the number in their name matches the order they exist in
the map and assigns each one an ID accordingly in the space object data.
This ID can then be used, once collision has been detected to trigger
some special action based on which checkpoint it was, e.g. a fight scene
or a sound recording.

For now we just print this number to the screen with the debug output as
a Proof of Concept (POC) to show that we are able to detect which
checkpoint we're at.