Collide zombies, player and dog using shapes instead of big squares
This works much better than before, you can precisely see when the two
shapes intersect and it's at a much closer range than when just using
the objects.  To get the shape to rotate about its centre instead of
origin (0,0) we need to either created it about origin, or recentre it
after creating it.  The unreleased version of resolve@master has a
helpful RecenterPoints method that does exactly this.  It needs to be
called on a ConvexPolygon though, so the Shapes we have need to be
converted to that explicit type before they can be used like this.

Also, to get the collision box to align with where the sprite is drawn
differently, the following changes were needed:
- centre the sprite around its centre instead of trying to
  post-manipulate the collision object's shape's position
- add some offset values I measured by hand for how far from the top the
  visual centre of the sprite is in the sprite sheet

This makes for pretty precise small squares that can be drawn around the
entities that move and rotate with them, are aligned visually to their
sprites and, after the broadphase collision check for cells, can be used
to check whether shapes within those cells are actually touching or not.