Shooting during reload interrupts it as dry firing
If you shoot during reload, the reload animation is instantly cancelled,
state is switched to dry fire, after which it will proceed back to idle.
The reload will not automatically restart, even if your ammo is empty,
so you can still shoot if you have some left but if this was an
automatic reload then the gun will still be empty and consecutive shots
will all dry fire until you manually reload.

Resolves #34