Fix player animation frame increment logic
I confess I don't understand what the previous code was supposed to.  I
logged it at different points along with its relate states and I don't
really understand how it was supposed to work.  The solution I've
replaced it with isn't a cool one-liner, it's a couple of boring if
statements but it works:

1. If the current frame is out of range of the current animation state
   then reset it to the first (from) frame
2. If we haven't reached the last (to) frame yet then increment the
   current animation frame

Fixes #13