Remove redundant animation logic
As I was attempting to find a solution the previous animation problem I
added several conditions and some of them have the same effect or are
checking conditions that are always true by the time the code gets that
far.  I have merged these superfluous conditions, deleted the rest and
the function is shorter and simpler for it.