Extract animation frame logic into a common function
All three entities player/dog/zombie had almost exactly the same
animation code.  I extracted the common part to animate.go:

> Given a current frame, a tick and a frameset, return the next frame

The animation-triggered logic for player and zombie that had crept into
their animate() functions was moved to their own function where they can
be handled concisely with a switch-case.  In the case of zombie this
required a bit of refactoring the Update() function to get the
end-states working correctly again because the order the animation and
the animation-triggered states were called in mattered, especially for
death.  In the case of the dog there was no such triggered logic but an
additional step was needed to use the state-to-frames map.