Add a HUD showing the amount of ammo you have left
There are 2 new images: Bullet, and Casing representing usable and spent
ammo in the gun.  Using playerAmmoClipMax and g.Player.Ammo it draws
these as icons at the bottom-right of the screen.  When a shot is fired
the used bullet is represented as an empty casing and when you finish
reloading the fill up.  This is all dynamic because those actions happen
elsewhere, this HUD code is just visualising it to the player.