Make death text red and fancy like-in another well-known game
The etxt library provides a more convenient API for handling text than
using Ebitengine directly, with this for example it was easy to change
the font size and experiment with it after the font was loaded.  I had
to convert the from from TTF to OTF otherwise it loaded with an error
about some outdated header data in it.

Now the text is big, red, centred on a black screen and uses the cool
font that looks similar to... the Dark Souls death screen.