Reduce debug print text
This was beginning to take up too much space on the screen so I've
deleted everything except FPS and Checkpoint.  FPS is important to have
on during development to see if a code change makes it drop because then
that needs focus to optimise.  Checkpoint is important to keep there to
help debug the issue we have right now with triggering voice clip
playbacks at the right checkpoints and tracking which one we're at.

Before this game gets published both of these will be removed from the
non-development version of the builds.