Move death screen related code out to deathscreen.go
This is gets out some death-specific stuff from main.go and because it's
used in the same way as startscreen.go a pattern is starting to emerge
that could be used to simplify these.

One thing I don't like is that because the text shown depends on some
game state like who died this is now set in the code of player/dog
depending on who died, this could definitely be improved because that's
not the right place for it for sure.