Extract all game logic to its own file
This took a while to refactor, mostly renaming things from Game to
GameScreen.  The state handling is a little better than before but it's
not quite done yet, because some extra logic will be needed to pass
information between the screens of the game states, for example right
now when you die, the death screen doesn't know if you died or the dog
because that dog's death state is in GameScreen, not DeathScreen.  How
much easier was life when it was all just in one big file...