Restart the game when you die
This calls the same NewGameScreen function from a go routine like the
main function does.   Because this function needs access to Game to be
able to change its state to running when it's done it now also has to
receive this tramp data to be able to pass it on.  We need a better way
of handling top-level state.  The audio context has been moved to the
top of the file with an init function because it panics if you try to
initialise it more than once and this is annoying.