Make sprinter zombies much faster
In the game these come sprinting at you from a distance but not with
such speed that you'd get as scared of them as their scream would lead
you to believe.  Now they'll be much faster than all the other zombies,
quite a bit faster than the player's normal running speed, but you can
still outrun it if you sprint.