Add script to generate animation frame tags from JSON
This script was written in a hurry and is probably brittle and not too
portable.  It assumes you have the `jq` command installed for parsing
the JSON.  It'll output an iota of frame tags with a provided prefix
into a new file of the given name.  The same prefix will also be used to
define a custom type (uint8) for these tags.  This script would probably
be more robust if it were written in Go but I was in a hurry right now.