Treat Boss type like Zombie
For this, the Zombies slice that's used to update zombies in the game
was modified to hold a slice of Zombielikes instead of concrete Zombies.
This way it's fine to put in both Boss and Zombie into it and it will
treat them the same.  No extra work is required on the Boss side because
its embedded Zombie type already implements everything that a Zombie
could do before.

Swapping out the concrete Zombie type for the Zombielike interface
required some additional work though, because some other parts of the
code (like gamescreen and dog) would directly reference fields like
Zombie.Object to do distance calculations and other code like spawnpoint
would reference the SpawnPoint or Space fields when removing zombies.
One way to solve this would've been to do a type assertion from
Zombielike to Zombie before all of these calls but then additional
conditionals would've been needed to handle the Boss, which is a
distinct type from Zombie.  Instead, the Zombielike interface now
includes 2 methods for this purpose:
- Position() returns a Coord of where the Zombie is
- Remove() removes the Zombie from its collision space and spawn point
This way all the previous call points could be rewritten to call these
instead of managing zombies using their fields from the outside.

As a side effect, the distance calculation function now takes Coords
instead of *resolv.Objects, which is okay because internally it was just
using coordinate float64s for the maths and no other object fields, so
only the function signature needed to be changed.  This change of
function signature however also meant that Player and Dog needed a way
to provide their position in the matching type Coord, so I gave them
Position() methods too.  In future it might be cool to generalise this.