Add animation state control for big boss
Reduced the boss hit points to 4 and set up 4 working animation states
that are tied to the inner zombies states.  Something is very strange
about how this is working because Animate gets called twice but it
didn't work nicely when I moved the inner Update code to the boss so I
kept it like this for now and instead override the inner animation frame
with the outter one.