Implement boss zombie phase two
Aka the daemon phase, this has 2 steps of death animation, from the big
one dying, to the small one spawning out of it.  The small one has extra
speed which I've hard-coded as the sprinter top-speed and takes on hit
to kill. So many exceptions and re-evaluations of the same states were
needed that it makes this code a hacky pile of workarounds around itself
instead of making it actually make sense.

I hope nobody ever reads this code but the deadline is nigh.