Fix boss zombie infinite death loop
When killing the boss zombie it had most of the states mapped including
death but not the sneaky non-animation state "dead" for after it has
died.  Instead of adding a fake state for this we're using the Dead
field (previously called ReallyDead out of frustration).  Like with the
inner zombie this will early-return an error and skip updating the
zombie, thereby avoiding the infinite loop.  This error is then handled
by the Zombies slice which will call Remove on this zombie so that it
can clean itself up before it is permanently removed from memory.

This meant reimplementing Die() and Remove() on the boss (some of which
is delegated) but Remove is critical here because the inner remove won't
find the zombie in the spawn point because the boss is registered there,
not the inner zombies.  Because the boss is its own type, not a regular
old Zombie, this also meant updating SpawnPoint.RemoveZombie to take the
Zombielike interface instead.