Synchronise changes to the top-level game state with a mutex
The go race detector found some race conditions that happen because of
the go routine running for NewGameScreen.  That function's very last
lines of code would set the top-level state as a signal that it is done
setting up and the Game can change to the next screen.  However while
that same code is running there are several pieces of logic in Update
which depend on the top-level state, and Draw is also using that same
state to determine which screen to draw and these do not necessarily run
in sync.  That means the value of state could be override even during
using it in e.g. Update, which is not safe.

To solve this I've added a R/W mutex StateLock at the top level.  Code
that modifies the State value like NewGameScreen and a small part of
Update code acquire a write lock before doing so.  Code that only reads
the state, like Draw, or some post-state-change triggers in Update only
acquire a read lock.  This differentiation between read and write locks
is necessary because if both NewGameScreen and Draw would acquire write
locks on the game state then they would have to wait for each other to
finish.  Between the two, NewGameScreen takes much longer, and pausing
Drawing until NewGameScreen is done would halt drawing and make it
pointless to have NewGameScreen loading assets in a go routine in the
background in the first place.