Show asset loading progress on the loading screen
This involves sharing a counter from the loading screen to the
NewGameScreen function.  NewGameScreen increments it every time it makes
an arbitrary amount of progress that I've classified by hand (luckily
they were already kinda grouped into sections).  The loading screen maps
these values to string representations that it displays to the user on
screen while they wait for the assets to load... if their computer is
slow enough to even see this screen for more than a split second or
they're using the web version.