Show the loading screen credits for a few seconds
Even if the game loads very fast show it for a few seconds, so whichever
one finishes last:

  * 2 seconds pass
  * all the assets load

The loading screen will wait it and then fade out to black.  This gives
someone a chance to glimpse the credits even if the assets load almost
instantly on a very fast computer and stop the screen from being a
jarring black flash at the beginning that looks like a bug.

This also presented a great opportunity to improve how the loading done
signal is sent from NewGameScreen back to the top-level game.  Instead
of changing its state directly (which never should've been in the
responsibility of NewGameScreen), once it's done, it sets Game.Loaded to
true.  The Game.Update method will then pass on this Loaded state to the
Loading screen so it can use it to see if it should still wait or it can
fade out already.  When the fading out itself is completed loading
screen will return with a "loading done" error that the top-level
Game.Update can use to decide which state it wants next, running or dead
(dead to reset to a later checkpoint) and exit early so that that state
is used instead of letting the error quit the game.