Increase boss hitpoints back from 7 to 10
This is much more exciting because you don't have enough bullets and
need to reload in between.  The animation sequence for hitpoints goes
like:

- 3 hits for left arm
- 3 hits for right arm
- 2 hits for torso
- 1 hit for demon

Hmmm that adds up to 9, so something is not quite how I thought it was
anyway it seems to kinda work like this in the game.