Simplify collision debugger
This was crashing trying to draw the boss zombie because it wants to get
an Object from it assuming it's a normal Zombie but for the boss that's
deeper down.  Instead of listing all the things to be drawn, just draw
all the things that exist in the collision space, that way I don't need
to add special logic for each one, it just works.