Generate audio players on the fly instead of rewinding
There was a performance issue with rewinding a pool of pre-loaded audio
players and also rewinding would interrupt the currently playing sound,
e.g. if there are many zombies and they both growl with the same growl
sound at the same time.

During loading the game we only load the bytes of the sound file into
memory instead of setting up a Vorbis stream in advance.  Then, when we
want to play the sound, we make a new Vorbis stream, make an audio
player out of it, play it and save a reference to it as the last played
sound in case it needs to be paused.  The ones that don't have a
reference to them any more are garbage collected once they're done
playing.  This solves the above interruption problem and so far seems
more performant than the previous solution.