Set different cursor modes for web and desktop versions
CursorModeCaptured doesn't seem to work on web, the cursor gets stuck at
wherever it was at the point the mouse leaves the screen, in fact it is
not really captured.  It works fine on desktop though, so using build
flags we'll now have 2 variations of cursor mode settings:
  - captured on desktop
  - hidden on web
Hopefully this keeps the good functionality on desktop while keeping it
at least somewhat playable on web, because right now it's impossible to
play on web.