Sync release version's FPS with vsync
For development it was useful to uncap FPS and let it update as much as
possible to be able to see with the impact code changes have on
performance, i.e. there's a much bigger drop if the FPS is uncapped.
For the release version though this should be in sync with vsync
otherwise it'd use too much power.