Put the full music back in the game
The shorter music track was used to keep the download file small during
development and testing but it was short enough that you'd hear the
tracks loop over.  Using the full version now because today is release
day.  Still keeping the short one in the repo because in future it might
be smart to use that for the test version and only use the full one for
the release version.