Fix the boss death-respawn state change
There are 2 things that affect state:
- end-frame based state changes
- hit-point based state changes

Here, after triggering the death animation from the hit-point based
changes it wants to continue to the next one but it's overridden by the
same hit-point based change.  I've moved all the hit-point-based stuff
to its own function and they're now only called if the boss isn't busy
dying, that way it'll only be the animation-based changes happening for
the whole duration of the death-respawn animation sequence.